requirements.txt
requests
beautifulsoup4
selenium
PyGithub
PyNaCl